backend/pyproject.toml
[project]
name = "diasense-backend"
version = "0.1.0"
description = "DiaSense FastAPI backend"
requires-python = ">=3.11,<3.14"
dependencies = [
  "fastapi>=0.115.6",
  "uvicorn[standard]>=0.34.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "pydantic[email]>=2.10.4",
  "pydantic-settings>=2.7.0",
  "python-multipart>=0.0.20",
  "pytesseract>=0.3.13",
  "Pillow>=11.1.0",
  "pypdf>=5.1.0",
  "pdf2image>=1.17.0",
  "joblib>=1.4.2",
  "numpy>=2.2.1",
  "pandas>=2.2.3",
  "prisma>=0.15.0",
  "scikit-learn>=1.6.0"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.8.6"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"
